+++ src/common/UITestAutomation/ModuleConfigData.cs
@@ -27,13 +27,7 @@ namespace Microsoft.PowerToys.UITest
     public enum PowerToysModule
     {
         PowerToysSettings,
-        FancyZone,
-        Hosts,
         Runner,
-        Workspaces,
-        PowerRename,
-        CommandPalette,
-        ScreenRuler,
         LightSwitch,
     }
 
@@ -100,13 +94,7 @@ namespace Microsoft.PowerToys.UITest
             ModuleInfo = new Dictionary<PowerToysModule, ModuleInfo>
             {
                 [PowerToysModule.PowerToysSettings] = new ModuleInfo("PowerToys.Settings.exe", "PowerToys Settings", "WinUI3Apps"),
-                [PowerToysModule.FancyZone] = new ModuleInfo("PowerToys.FancyZonesEditor.exe", "FancyZones Layout"),
-                [PowerToysModule.Hosts] = new ModuleInfo("PowerToys.Hosts.exe", "Hosts File Editor", "WinUI3Apps"),
                 [PowerToysModule.Runner] = new ModuleInfo("PowerToys.exe", "PowerToys"),
-                [PowerToysModule.Workspaces] = new ModuleInfo("PowerToys.WorkspacesEditor.exe", "Workspaces Editor"),
-                [PowerToysModule.PowerRename] = new ModuleInfo("PowerToys.PowerRename.exe", "PowerRename", "WinUI3Apps"),
-                [PowerToysModule.CommandPalette] = new ModuleInfo("Microsoft.CmdPal.UI.exe", "PowerToys Command Palette", "WinUI3Apps\\CmdPal"),
-                [PowerToysModule.ScreenRuler] = new ModuleInfo("PowerToys.MeasureToolUI.exe", "PowerToys.ScreenRuler", "WinUI3Apps"),
                 [PowerToysModule.LightSwitch] = new ModuleInfo("PowerToys.LightSwitch.exe", "PowerToys.LightSwitch", "LightSwitchService"),
             };
         }